- Refactoring - deleted UnitOfWork classes - Created SqliteDaoBase class and modified all DaoImpl classes to extend from the base class - Created 2 new packages and moved classes as required - Renamed Interface classes with prefix I - standardized DAO update methods by removing column-specific update methods - Removed un-Updatabale columns from DAO update methods - some minor changes like removing commented code

+++ src/com/theAlternate/lifrack/Reminder.java
@@ -13,7 +13,7 @@ import android.util.Log;
 
 
 public class Reminder implements Parcelable{
-	private static final int DUMMY_VALUE = -1;
+	private static final long DUMMY_VALUE = -1;
 	private long habitId;
 	private long id;
 	private int hour;
